+++ src/main/java/com/yesipchuk/demo/DAO/ingredients/impls/IngredientsDAOFakeImpl.java
@@ -2,7 +2,7 @@ package com.yesipchuk.demo.DAO.ingredients.impls;
 
 import com.yesipchuk.demo.DAO.ingredients.interfaces.IIngredientsDao;
 import com.yesipchuk.demo.datastorage.DataStorageJDBC;
-import com.yesipchuk.demo.modelJDBC.Ingredients;
+import com.yesipchuk.demo.model.Ingredients;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -23,10 +23,10 @@ public class IngredientsDAOFakeImpl implements IIngredientsDao {
         ResultSet resultSet;
         resultSet = dataStorage.executeQuery("SELECT * FROM ingredients");
         while(resultSet.next()){
-            list.add(new Ingredients((long)resultSet.getInt("id"),
+            list.add(new Ingredients(resultSet.getInt("id"),
                     resultSet.getString("name"),
-                    (long)resultSet.getInt("critical_rate"),
-                    (long)resultSet.getInt("amount"),
+                    resultSet.getInt("critical_rate"),
+                    resultSet.getInt("amount"),
                     resultSet.getDouble("price")));
         }
         return list;
